Support type annotations for class consts (fixes #942).

+++ src/Psalm/Issue/CodeIssue.php
@@ -10,7 +10,9 @@ use function explode;
 
 abstract class CodeIssue
 {
+    /** @var int */
     public const ERROR_LEVEL = -1;
+    /** @var int<0, max> */
     public const SHORTCODE = 0;
 
     /**
@@ -88,8 +90,8 @@ abstract class CodeIssue
             $snippet_bounds[1],
             $location->getColumn(),
             $location->getEndColumn(),
-            (int) static::SHORTCODE,
-            (int) static::ERROR_LEVEL,
+            static::SHORTCODE,
+            static::ERROR_LEVEL,
             $this instanceof TaintedInput
                 ? $this->getTaintTrace()
                 : null,
